requirements.txt
pandas>=2.2
numpy>=2.0
PyYAML>=6.0
pyarrow>=16.0
pytest>=8.0

